+++ nextjs-app/shared/components/HelperText/index.ts
@@ -1,2 +1,2 @@
 export { default } from "./HelperText";
-export type { HelperTextProps } from "./HelperText";
+export type { HelperTextProps, HelperTextState } from "./HelperText";